% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions Organiza informacoes de manutencao.R
\name{Arla_trocas}
\alias{Arla_trocas}
\title{Arla Trocas}
\usage{
Arla_trocas(aux_df)
}
\arguments{
\item{aux_df}{the data.frame containing the columns "TEXT" from which the informations will be extracted.}

\item{add_patt}{if \code{TRUE} a column will be added informating what pattern was used to extrach the informations}
}
\value{
The same \code{data.frame} containing extra columns and the respective values extraced from the "TEXT" column.
}
\description{
Extract the information of Arla replacements
}
\examples{

}
