% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions Organiza informacoes de manutencao.R
\name{ArMotor_trocas}
\alias{ArMotor_trocas}
\title{Ar Motor Trocas}
\usage{
ArMotor_trocas(aux_df, add_patt = TRUE)
}
\arguments{
\item{aux_df}{the data.frame containing the columns "TEXT" from which the informations will be extracted.}

\item{add_patt}{if \code{TRUE} a column will be added informating what pattern was used to extrach the informations}
}
\value{
The same \code{data.frame} containing extra columns and the respective values extraced from the "TEXT" column.
}
\description{
Extract the information of Air Engine oil replacements
}
\examples{

}
