% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions Organiza informacoes de manutencao.R
\name{f_DE_PARA}
\alias{f_DE_PARA}
\title{Função DE-PARA}
\usage{
f_DE_PARA(aux.da, aux.DE_PARA)
}
\arguments{
\item{aux.da}{A vector containing the data to have the replacements}

\item{aux.DE_PARA}{A data.frame with the "DE PARA" rules}
}
\value{
o mesmo vetor com os valores modificados
}
\description{
A função muda os valores do vetor de acordo com a abela de-para fornecida
}
\details{
Sem detalhes
}
\examples{

}
