% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions Organiza informacoes de manutencao.R
\name{f_primeira_troca_vira_segunda}
\alias{f_primeira_troca_vira_segunda}
\title{First replacement becomes the second.}
\usage{
f_primeira_troca_vira_segunda(aux.da, col_name)
}
\arguments{
\item{aux.da}{a data frame}

\item{col_name}{a string containing the name of the columns to have the value replaced}
}
\description{
switches values of columns col_name and col_name2 when the value of in col_name is NA
}
\examples{

}
