% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions Organiza informacoes de manutencao.R
\name{mileage_correction}
\alias{mileage_correction}
\title{Correção de quilometragem}
\usage{
mileage_correction(aux_df)
}
\arguments{
\item{aux_df}{the \code{data.frame} to be modified}
}
\description{
Multiplies the mileage according to the column "service_class".
}
\examples{

}
