% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions Organiza informacoes de manutencao.R
\name{motor_trocas}
\alias{motor_trocas}
\title{Motor Trocas}
\usage{
motor_trocas(aux_df, add_patt = FALSE)
}
\arguments{
\item{aux_df}{the data.frame containing the columns "TEXT" from which the informations will be extracted.}

\item{add_patt}{if \code{TRUE} a column will be added informating what pattern was used to extrach the informations}
}
\value{
The same \code{data.frame} containing 3 extra columns and the respective values extraced from the text column.
}
\description{
Extract the information of Engine oil replacements
}
\examples{

}
